requirements.txt
aiohttp
python-decouple
python-bale-bot
